// src/api/generated/schemas/audienceGroupDto.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */
import type { AudienceType } from './audienceType';
import type { AudienceFilterRule } from './audienceFilterRule';

export interface AudienceGroupDto {
  id?: string;
  organizationId?: string;
  name?: string;
  /** @nullable */
  description?: string | null;
  type?: AudienceType;
  /** @nullable */
  filterExpressions?: AudienceFilterRule[] | null;
}