// src/api/generated/schemas/updateAudienceRequest.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */
import type { AudienceType } from './audienceType';
import type { AudienceFilterRule } from './audienceFilterRule';

export interface UpdateAudienceRequest {
  name?: string;
  /** @nullable */
  description?: string | null;
  type?: AudienceType;
  /** @nullable */
  filterExpressions?: AudienceFilterRule[] | null;
}